Feature/api테스트

See merge request s13-webmobile2-sub1/S13P11A109!21

+++ Ticket_BackEnd/src/main/java/com/ssafy/ticket_backend/dto/request/CommentRequest.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class CommentRequest {
 
-    int userId;      // TODO 시큐리티
+    Long userId;
     Long postId;
     String content;
 }
